requirements.txt
pyocr
pillow
opencv-python
wand
